Move functions to lib.rs from main
